+++ enershelf/map/config.py
@@ -36,7 +36,7 @@ FILTER_DEFINITION = {}
 
 # STORE
 
-STORE_COLD_INIT = json.dumps({"version": __version__})
+STORE_COLD_INIT = {"version": __version__}
 
 
 def init_hot_store():